chore: split object method and object property out
