+++ object.go
@@ -45,7 +45,7 @@ func (df DotFunction) Calculate(val entry, cfg *treeConfig) entry {
 		return rhsVal
 	}
 
-	return vFv // this will be an Undefined type.
+	return vFv // this will already be an Undefined type.
 }
 
 type Member interface{ Variable }
@@ -72,84 +72,6 @@ type ObjectValue struct {
 
 func (o ObjectValue) String() string {
 	return fmt.Sprintf("ObjectValue(%T)", o.Object)
-}
-
-// TODO: could we use the same principle as Function.Receiver with Variable? Would it be elegant?
-// ObjectProperty is a Tree entry that holds a reference of a user-defined object by name and the property to access on it.
-// It is used to access a property on a user-defined object.
-// It is a "cousin" of Variable, but for a property of a user-defined object.
-type ObjectProperty struct {
-	ObjectName   string
-	PropertyName string
-}
-
-func NewObjectProperty(objectName, propertyName string) ObjectProperty {
-	return ObjectProperty{
-		ObjectName:   objectName,
-		PropertyName: propertyName,
-	}
-}
-
-//nolint:errcheck // life's too short to check for type assertion success here
-func (o ObjectProperty) Calculate(val entry, op Operator, cfg *treeConfig) entry {
-	rhsVal := cfg.ObjectProperty(o)
-	if u, ok := rhsVal.(Undefined); ok {
-		return u
-	}
-
-	if val == nil {
-		return rhsVal
-	}
-
-	val = calculate(val.(Value), op, rhsVal)
-
-	return val
-}
-
-func (o ObjectProperty) String() string {
-	return fmt.Sprintf("%s.%s", o.ObjectName, o.PropertyName)
-}
-
-// ObjectMethod is a Tree entry that holds a reference of a user-defined object by name and the method to call on it.
-// It is used to call a method on a user-defined object.
-// It is a "cousin" of Function, but for a method of a user-defined object.
-type ObjectMethod struct {
-	ObjectName string
-	MethodName string
-	Args       []Tree
-}
-
-func NewObjectMethod(objectName, propertyName string, args ...Tree) ObjectMethod {
-	return ObjectMethod{
-		ObjectName: objectName,
-		MethodName: propertyName,
-		Args:       args,
-	}
-}
-
-//nolint:errcheck // life's too short to check for type assertion success here
-func (om ObjectMethod) Calculate(val entry, op Operator, cfg *treeConfig) entry {
-	// attempt to get body of a user-provided object's method.
-	bodyFn := cfg.ObjectMethod(om)
-
-	fn := NewFunction(om.MethodName, bodyFn, om.Args...)
-
-	rhsVal := fn.Eval(WithFunctions(cfg.functions), WithVariables(cfg.variables), WithObjects(cfg.objects))
-	if u, ok := rhsVal.(Undefined); ok {
-		return u
-	}
-
-	if val == nil {
-		return rhsVal
-	}
-
-	val = calculate(val.(Value), op, rhsVal)
-
-	return val
-}
-
-func (om ObjectMethod) String() string {
-	return fmt.Sprintf("%s.%s", om.ObjectName, om.MethodName)
 }
 
 // ObjectGetProperty returns the value of the property with the given name from the object.
